version

+++ files/CryptionGUI (ALPHA) PY for exe/CryptionGUI.py
@@ -107,18 +107,22 @@ def restart(eordb, e, restartb, exitb, clipL):
 
 def version_check():
 	url = "https://raw.githubusercontent.com/Agwebberley/agwebberley.tk/master/files/newestversion.json"
+	url2 = "https://raw.githubusercontent.com/Agwebberley/agwebberley.tk/master/files/newestversionexe.json"
 	where = os.getcwd()
 	where = str(where)
 	wget.download(url, where)
+	wget.download(url2, where)
 
 	with open('newestversion.json') as json_file:
 		newestversion = json.load(json_file)
 	with open('version.json') as json_file:
 		version = json.load(json_file)
+	with open('newestversionexe.json') as json_file:
+		newestversionexe = json.load(json_file)
 
-	if version != newestversion:
-		neednewL = Label(root, text="A new version is availible")
-		neednewL.pack()
+	if version != newestversionexe:
+		neednewexeL = Label(root, text="New version availible for exe")
+		neednewexeL.pack()
 		clipL = Label(root, text="placement")
 		restartb = Button(root, text="Continue Anyway", command=lambda: restart(eordb, e, restartb, exitb, clipL))
 		restartb.pack()
@@ -126,8 +130,27 @@ def version_check():
 		exitb.pack()
 		if os.path.exists("newestversion.json"):
 			os.remove("newestversion.json")
-	if os.path.exists("newestversion.json"):
-		os.remove("newestversion.json")
+		if os.path.exists("newestversionexe.json"):
+			os.remove("newestversionexe.json")
+	else:
+
+		if version != newestversion:
+			neednewL = Label(root, text="A new version is availible for py")
+			neednewL.pack()
+			clipL = Label(root, text="placement")
+			restartb = Button(root, text="Continue Anyway", command=lambda: restart(eordb, e, restartb, exitb, clipL))
+			restartb.pack()
+			exitb = Button(root, text="Quit", command=root.quit)
+			exitb.pack()
+			if os.path.exists("newestversion.json"):
+				os.remove("newestversion.json")
+			if os.path.exists("newestversionexe.json"):
+				os.remove("newestversionexe.json")
+
+		if os.path.exists("newestversion.json"):
+			os.remove("newestversion.json")
+		if os.path.exists("newestversionexe.json"):
+			os.remove("newestversionexe.json")
